+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Remove landing domain and guides content (DOP-2215, #314)
+
 ## [v0.9.9] - 2021-06-07
 
 ### Added